+++ database/factories/UserFactory.php
@@ -1,6 +1,7 @@
 <?php
 
 use Faker\Generator as Faker;
+use Illuminate\Support\Facades\Hash;
 
 /*
 |--------------------------------------------------------------------------
@@ -15,9 +16,13 @@ use Faker\Generator as Faker;
 
 $factory->define(App\User::class, function (Faker $faker) {
     return [
-        'name' => $faker->name,
-        'email' => $faker->unique()->safeEmail,
-        'password' => '$2y$10$TKh8H1.PfQx37YgCzwiKb.KjNyWgaHb9cbcoQgdIVFlYg7B77UdFm', // secret
+        'name'           => $faker->name,
+        'department'     => $faker->name,
+        'email'          => $faker->unique()->safeEmail,
+        'phone'          => $faker->phoneNumber,
+        'password'       => Hash::make($faker->numberBetween($min = 1000, $max = 9000)),
+        'role'           => array_random(['ADMIN', 'EMPLOYEE']),
         'remember_token' => str_random(10),
     ];
+
 });