+++ database/migrations/2018_05_23_092726_create_borrows_table.php
@@ -4,7 +4,7 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class CreateUsersTable extends Migration
+class CreateBorrowsTable extends Migration
 {
     /**
      * Run the migrations.
@@ -13,15 +13,12 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) {
+        Schema::create('borrows', function (Blueprint $table) {
             $table->increments('id');
-            $table->string('name');
-            $table->string('department');
-            $table->string('email')->unique();
-            $table->string('phone');
-            $table->string('role');
-            $table->string('password');
-            $table->rememberToken();
+            $table->string('employeeId');
+            $table->string('category');
+            $table->string('itemName');
+            $table->string('urgency');
             $table->timestamps();
         });
     }
@@ -33,6 +30,6 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('users');
+        Schema::dropIfExists('borrows');
     }
 }