+++ resources/views/layouts/dash.blade.php
@@ -1,0 +1,111 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <!-- Tell the browser to be responsive to screen width -->
+    <meta content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" name="viewport">
+
+    <!-- CSRF Token -->
+    <meta name="csrf-token" content="{{ csrf_token() }}">
+
+    <title>{{ config('app.name') }}</title>
+
+    <link rel="stylesheet" href="{{ asset('css/dash.css') }}">
+    <link rel="stylesheet" href="{{ asset('css/font-awesome/css/font-awesome.min.css') }}">
+    <link rel="stylesheet" href="{{ asset('css/ionicons/dist/css/ionicons.min.css') }}">
+
+    <!-- Google Font -->
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,600,700,300italic,400italic,600italic">
+</head>
+<body class="hold-transition skin-blue sidebar-mini">
+<!-- Site wrapper -->
+<div class="wrapper">
+
+@include('shared.dash.header')
+<!-- =============================================== -->
+
+    <!-- Left side column. contains the sidebar -->
+@include('shared.dash.sidebar')
+
+<!-- =============================================== -->
+
+    <!-- Content Wrapper. Contains page content -->
+    <div class="content-wrapper">
+        <!-- Content Header (Page header) -->
+        <section class="content-header">
+            @yield('page-title')
+        </section>
+
+        <!-- Main content -->
+        <section class="content">
+
+            <!-- Default box -->
+            <div class="box">
+
+                <div class="box-header with-border">
+                    @yield('content-header')
+                </div>
+
+                <div class="box-body">
+                    @yield('content')
+                </div>
+                <!-- /.box-body -->
+
+                <div class="box-footer">
+                    @yield('content-footer')
+                </div>
+                <!-- /.box-footer-->
+            </div>
+            <!-- /.box -->
+
+        </section>
+        <!-- /.content -->
+    </div>
+    <!-- /.content-wrapper -->
+
+@include('shared.dash.footer')
+
+<!-- Control Sidebar -->
+@include('shared.dash.control')
+<!-- /.control-sidebar -->
+    <!-- Add the sidebar's background. This div must be placed
+         immediately after the control sidebar -->
+    <div class="control-sidebar-bg"></div>
+</div>
+<!-- ./wrapper -->
+
+<form id="logout-form" action="{{ route('logout') }}" method="POST" style="display: none;">
+    @csrf
+</form>
+
+<script src="{{ asset('js/dash.js') }}"></script>
+
+
+@stack('footer-scripts')
+
+<script>
+    $(function () {
+        $('.sidebar-menu').tree();
+
+        // Submit logout form
+        $('.sign-out').click(function () {
+            $('#logout-form').submit()
+        });
+
+        // CSRF protection for AJAX calls
+        $.ajaxSetup({
+            headers: {
+                'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
+            }
+        });
+
+        $('#datetimepicker').datetimepicker({
+            format: 'dd/mm/yyyy'
+        });
+
+
+    })
+</script>
+</body>
+</html>