+++ CHANGELOG.md
@@ -3,15 +3,28 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.1.4
+###Added
+* Pdf generator
+
+
+## 0.1.3
+###Added
+* Asset request
+* Requested assets
+
 ## 0.1.2
 ### Added
-
+* Organization assets
+* Asset distribution
 
 ## 0.1.1 - 2018-05-22
 ### Added
+* User login
+* User seeders
 
 
-## 0.1.0 - 2018-05-18
+## 0.1.0 - 2018-05-22
 ### Added
 * Changelog
 * ReadMe file